+++ src/components/Setup/SetupLogin.tsx
@@ -1,6 +1,6 @@
 import Button from '@app/components/Common/Button';
-import JellyfinLogin from '@app/components/Login/JellyfinLogin';
-import PlexLoginButton from '@app/components/PlexLoginButton';
+import PlexLoginButton from '@app/components/Login/PlexLoginButton';
+import JellyfinSetup from '@app/components/Setup/JellyfinSetup';
 import { useUser } from '@app/hooks/useUser';
 import defineMessages from '@app/utils/defineMessages';
 import { MediaServerType } from '@server/constants/server';
@@ -83,11 +83,9 @@ const SetupLogin: React.FC<LoginWithMediaServerProps> = ({
       </div>
       {serverType === MediaServerType.PLEX && (
         <>
-          <div
-            className="px-10 py-8"
-            style={{ backgroundColor: 'rgba(0,0,0,0.3)' }}
-          >
+          <div className="flex justify-center bg-black/30 px-10 py-8">
             <PlexLoginButton
+              large
               onAuthToken={(authToken) => {
                 setMediaServerType(MediaServerType.PLEX);
                 setAuthToken(authToken);
@@ -102,16 +100,14 @@ const SetupLogin: React.FC<LoginWithMediaServerProps> = ({
         </>
       )}
       {serverType === MediaServerType.JELLYFIN && (
-        <JellyfinLogin
-          initial={true}
+        <JellyfinSetup
           revalidate={revalidate}
           serverType={serverType}
           onCancel={onCancel}
         />
       )}
       {serverType === MediaServerType.EMBY && (
-        <JellyfinLogin
-          initial={true}
+        <JellyfinSetup
           revalidate={revalidate}
           serverType={serverType}
           onCancel={onCancel}
